pyproject.toml
[project]
name = "imagerecognition"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "docutils>=0.22.4",
    "flask>=3.1.2",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "sphinx>=9.1.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
]
